test: decompose suite by semantic owner
