+++ tests/analysis/test_scaling_recipes.py
@@ -2,36 +2,12 @@ from __future__ import annotations
 
 from datetime import UTC, datetime
 from pathlib import Path
-from typing import NoReturn
-
-import pytest
 
 from flameox.analysis import RecipeService
 from flameox.catalog import Catalog
-from flameox.domain import DomainError, ErrorCode, canonical_json
+from flameox.domain import canonical_json
 from flameox.evidence import GenerationPublisher
 from flameox.storage import Workspace
-
-
-def _run_row(run_id: str) -> dict[str, object]:
-    return {
-        "run_id": run_id,
-        "created_at": datetime.now(UTC),
-        "run_type": "execution",
-        "execution_status": "succeeded",
-        "capture_status": "complete",
-        "validation_status": "passed",
-        "workload_definition_id": "workload",
-        "workload_instance_id": "workload-instance",
-        "measurement_protocol_id": "protocol",
-        "environment_id": "environment",
-        "source_state_id": "source",
-        "collector": "fixture",
-        "collector_version": "1",
-        "exit_code": 0,
-        "wall_time_ns": 1,
-        "manifest_path": f"runs/{run_id}/manifest.json",
-    }
 
 
 def test_scaling_reports_dispersion_models_and_supported_range(
@@ -212,185 +188,6 @@ def test_scaling_reports_dispersion_models_and_supported_range(
     assert result.correlated_hotspots[0].multiplicity_method == "benjamini-hochberg-fdr"
     assert result.correlated_hotspots[0].tested_hypothesis_count == len(result.correlated_hotspots)
     assert result.correlated_hotspots[0].adjusted_p_value >= result.correlated_hotspots[0].p_value
-
-
-def test_memory_reports_phase_correlated_growth(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    rows: list[dict[str, object]] = []
-    for index, (phase, value) in enumerate(
-        (("warmup", 100), ("steady_state", 240), ("shutdown", 260))
-    ):
-        rows.append(
-            {
-                "measurement_id": f"memory-{index}",
-                "run_id": "memory-run",
-                "artifact_id": None,
-                "name": "memory.retained_end",
-                "value_int": value,
-                "value_float": None,
-                "unit": "bytes",
-                "aggregation": "single",
-                "scope": "process",
-                "trial_id": None,
-                "worker_id": None,
-                "worker_run_index": 0,
-                "value_index": index,
-                "loop_count": None,
-                "is_warmup": phase == "warmup",
-                "block_id": None,
-                "variant_id": None,
-                "order_in_block": None,
-                "phase": phase,
-                "dimensions": {},
-                "evidence_level": "observed",
-            }
-        )
-    GenerationPublisher(workspace).publish_rows(
-        {"runs": [_run_row("memory-run")], "measurements": rows},
-        publisher="memory-fixture",
-        publisher_version="1",
-    )
-
-    result = RecipeService(workspace).memory("memory-run")
-
-    assert [point.phase for point in result.phase_growth] == [
-        "warmup",
-        "steady_state",
-        "shutdown",
-    ]
-    assert [point.delta for point in result.phase_growth] == [None, 140.0, 20.0]
-
-
-def test_execution_compares_path_and_semantic_observation_changes(
-    tmp_path: Path,
-) -> None:
-    workspace = Workspace.initialize(tmp_path)
-
-    def observation(
-        run_id: str,
-        observation_id: str,
-        name: str,
-        value: str,
-        line: int,
-    ) -> dict[str, object]:
-        return {
-            "observation_id": observation_id,
-            "run_id": run_id,
-            "artifact_id": None,
-            "kind": "configuration",
-            "name": name,
-            "value_json": value,
-            "file": "policy.py",
-            "line_from": line,
-            "line_to": line,
-            "context": "update",
-            "evidence_level": "observed",
-        }
-
-    GenerationPublisher(workspace).publish_rows(
-        {
-            "runs": [_run_row("baseline"), _run_row("candidate")],
-            "observations": [
-                observation("baseline", "old-source", "old_log_prob_source", '"rollout"', 10),
-                observation("baseline", "removed", "legacy_branch", "true", 20),
-                observation("candidate", "new-source", "old_log_prob_source", '"epoch"', 10),
-                observation("candidate", "added", "clip_fraction", "0.21", 30),
-            ],
-        },
-        publisher="execution-fixture",
-        publisher_version="1",
-    )
-
-    result = RecipeService(workspace).execution(
-        "baseline",
-        comparison_input_id="candidate",
-    )
-
-    assert [item.name for item in result.added] == ["clip_fraction"]
-    assert [item.name for item in result.removed] == ["legacy_branch"]
-    assert [item.name for item in result.changed] == ["old_log_prob_source"]
-    assert result.changed[0].baseline_value_json == '"rollout"'
-    assert result.changed[0].candidate_value_json == '"epoch"'
-
-
-def test_execution_comparison_diffs_complete_inputs_before_limiting(
-    tmp_path: Path,
-) -> None:
-    workspace = Workspace.initialize(tmp_path)
-
-    def observation(
-        run_id: str,
-        observation_id: str,
-        name: str,
-        value: str,
-        line: int,
-    ) -> dict[str, object]:
-        return {
-            "observation_id": observation_id,
-            "run_id": run_id,
-            "artifact_id": None,
-            "kind": "configuration",
-            "name": name,
-            "value_json": value,
-            "file": "policy.py",
-            "line_from": line,
-            "line_to": line,
-            "context": "update",
-            "evidence_level": "observed",
-        }
-
-    GenerationPublisher(workspace).publish_rows(
-        {
-            "runs": [_run_row("baseline"), _run_row("candidate")],
-            "observations": [
-                observation("baseline", "shared-old", "shared", '"old"', 10),
-                observation("baseline", "baseline-only", "baseline_only", "true", 20),
-                observation("candidate", "candidate-early", "candidate_early", "true", 1),
-                observation("candidate", "shared-new", "shared", '"new"', 10),
-            ],
-        },
-        publisher="bounded-execution-fixture",
-        publisher_version="1",
-    )
-
-    result = RecipeService(workspace).execution(
-        "baseline",
-        comparison_input_id="candidate",
-        limit=1,
-    )
-
-    assert [item.name for item in result.added] == ["candidate_early"]
-    assert [item.name for item in result.removed] == ["baseline_only"]
-    assert [item.name for item in result.changed] == ["shared"]
-    assert result.returned == 1
-    assert result.truncated
-
-
-def test_read_only_analysis_pins_snapshot_without_workspace_write_lock(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    head = workspace.corpus.read_head().commit_id
-
-    def fail_write_lock(*_args: object, **_kwargs: object) -> NoReturn:
-        raise AssertionError("read-only analysis acquired the workspace write lock")
-
-    monkeypatch.setattr(workspace, "write_locked", fail_write_lock)
-
-    result = RecipeService(workspace).failures(limit=10)
-
-    assert result.corpus_commit_id == head
-    assert workspace.corpus.read_head().commit_id == head
-
-
-def test_analysis_rejects_input_absent_from_pinned_corpus(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-
-    with pytest.raises(DomainError) as error:
-        RecipeService(workspace).hotspots("missing-run")
-
-    assert error.value.code is ErrorCode.WORKSPACE_INVALID
 
 
 def test_scaling_correlated_hotspots_empty_when_all_filtered(tmp_path: Path) -> None:
