+++ tests/support/__init__.py
@@ -1,0 +1,1 @@
+"""Narrow test builders and readiness helpers."""
